# SWATCH-2935: Improve find guests DB query