--- a/swatch-core/src/main/java/org/candlepin/subscriptions/db/HostRepository.java
+++ b/swatch-core/src/main/java/org/candlepin/subscriptions/db/HostRepository.java
@@ -527,7 +527,7 @@ public interface HostRepository
       "select distinct h1 from Host h1 where "
           + "h1.orgId = :orgId and "
           + "h1.hypervisorUuid in (select h2.subscriptionManagerId from Host h2 where "
-          + "h2.instanceId = :instanceId)")
+          + "h2.orgId=:orgId and h2.instanceId = :instanceId)")
   Page<Host> getGuestHostsByHypervisorInstanceId(
       @Param("orgId") String orgId, @Param("instanceId") String instanceId, Pageable pageable);